backend/requirements.txt
fastapi
uvicorn
easyocr
pillow
opencv-python
pydantic
langdetect
spacy
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz
google-cloud-vision
pypdfium2
